Rename ingestors base class
