+++ ingestors/documents/djvu.py
@@ -1,4 +1,4 @@
-from ingestors.base import Ingestor
+from ingestors.ingestor import Ingestor
 from ingestors.support.pdf import PDFSupport
 from ingestors.util import join_path
