+++ ingestors/media/audio.py
@@ -1,7 +1,7 @@
 import logging
 from pymediainfo import MediaInfo
 
-from ingestors.base import Ingestor
+from ingestors.ingestor import Ingestor
 from ingestors.media.util import MediaInfoDateMixIn
 
 log = logging.getLogger(__name__)
