+++ ingestors/media/svg.py
@@ -1,6 +1,6 @@
 import logging
 
-from ingestors.base import Ingestor
+from ingestors.ingestor import Ingestor
 from ingestors.support.html import HTMLSupport
 from ingestors.support.encoding import EncodingSupport
