+++ ingestors/packages/__init__.py
@@ -3,7 +3,7 @@ import gzip
 import shutil
 import tarfile
 
-from ingestors.base import Ingestor
+from ingestors.ingestor import Ingestor
 from ingestors.support.shell import ShellSupport
 from ingestors.support.package import PackageSupport
 from ingestors.exc import ProcessingException
