+++ ingestors/packages/zip.py
@@ -1,7 +1,7 @@
 import logging
 import zipfile
 
-from ingestors.base import Ingestor
+from ingestors.ingestor import Ingestor
 from ingestors.support.package import PackageSupport
 from ingestors.exc import ProcessingException
