+++ ingestors/tabular/csv.py
@@ -5,7 +5,7 @@ import csv
 import logging
 from collections import OrderedDict
 
-from ingestors.base import Ingestor
+from ingestors.ingestor import Ingestor
 from ingestors.support.encoding import EncodingSupport
 from ingestors.exc import ProcessingException
 from ingestors.util import safe_string
